pyproject.toml
[project]
name = "t-music-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "davey>=0.1.5",
    "discord-py>=2.7.1",
    "pynacl>=1.6.2",
    "python-dotenv>=1.2.2",
]
